Refactor various files: improve type casting, enhance function signatures, and streamline error handling

+++ src/config/ConfigManager.cpp
@@ -50,7 +50,7 @@ std::string ConfigManager::computeBlakeHash( const filesystem::path& filePath ) 
             throw std::runtime_error( "Error reading file during hashing: " + filePath.string() );
         }
         if ( file.gcount() > 0 ) {
-            if ( EVP_DigestUpdate( ctx, buf, file.gcount() ) != 1 ) {
+            if ( EVP_DigestUpdate( ctx, buf, (size_t) file.gcount() ) != 1 ) {
                 EVP_MD_CTX_free( ctx );
                 throw std::runtime_error( "EVP_DigestUpdate failed" );
             }
